[lldb][NFC] Move helpers to import record layout into ClangASTImporter
